requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fsspec==2025.5.1
gunicorn==23.0.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
numpy
onnx==1.18.0
onnxruntime
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
requests==2.32.4
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
